api/requirements-minimal.txt
Django==5.2.3
django-cors-headers==4.7.0
django-sendgrid-v5==1.3.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.4
sendgrid==6.12.4
sendgrid-django==4.2.0
boto3==1.38.40
botocore==1.38.40
s3transfer==0.13.0
